requirements.txt
Celery
Django
django-polymorphic
djangorestframework
psycopg2-binary
redis
whitenoise==5.2.0